+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Fixed
+- `FallbackMediaFacadeDecorator` no longer search for an empty fallback media id and trhows the original exception instead.
+
 ## [5.1.0] - 2026-05-13
 
 ### Added